requirements/base.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
dj-database-url>=2.0.0
psycopg2-binary
python-dotenv>=1.0.0
whitenoise
gunicorn
requests>=2.31.0
PyJWT>=2.8.0
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
django-mptt>=0.16.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-barcode>=0.15.1
qrcode>=7.4.0
